work/requirements.txt
nltk
pandas
numpy
matplotlib
pyLDAvis
kafka-python
requests